+++ DietApp/AppDelegate.swift
@@ -29,7 +29,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     return .portrait // デフォルトは縦向き
   }
-  
+
   func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
     let config = Realm.Configuration(
       schemaVersion: 1,
@@ -69,10 +69,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 extension AppDelegate: UNUserNotificationCenterDelegate {
   
-  // 通知に対するユーザーのアクションを処理するデリゲートメソッド
+  // ユーザーが通知に対してアクションをとった時に呼ばれるデリゲートメソッド
   // center: 通知を管理するUNUserNotificationCenterのインスタンス
   // response: ユーザーの応答情報を含むオブジェクト
   // completionHandler: 処理完了時に必ず呼び出す必要があるクロージャー
+  
   func userNotificationCenter(
     _ center: UNUserNotificationCenter,
     didReceive response: UNNotificationResponse,
@@ -81,14 +82,14 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
     // ユーザーが選択したアクションのタイプに基づいて処理を分岐
     switch response.actionIdentifier {
     case "RECORD_ACTION":  // レコードアクションが選択された場合
-      navigateToTopScreen()
+      LocalNotificationManager.shared.navigateToTopScreen()
     case "LATER_ACTION":  // 後で見るアクションが選択された場合
       print("Later action selected")
       break  // 特に追加の処理は必要なし
       
       //通知自体をタップした時の処理
     default:
-      navigateToTopScreen()
+      LocalNotificationManager.shared.navigateToTopScreen()
       break
     }
     // 処理完了をシステムに通知
@@ -96,6 +97,7 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
     completionHandler()
   }
   
+  //フォアグラウンドで通知を受信した時の処理
   func userNotificationCenter(
     _ center: UNUserNotificationCenter,
     willPresent notification: UNNotification,
@@ -105,43 +107,3 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
     completionHandler([.banner])
   }
 }
-//ボタンや通知自体をタップした時にアプリのトップ画面に飛ばす処理
-private func navigateToTopScreen() {
-  if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-     let window = windowScene.windows.first {
-    
-    // ルートビューコントローラーがUITabBarControllerであることを確認
-    guard let tabBarController = window.rootViewController as? UITabBarController else {
-      print("Error: Root view controller is not UITabBarController")
-      return
-    }
-    
-    // 目的の画面があるタブを選択（通常は最初のタブ）
-    tabBarController.selectedIndex = 0
-    
-    // 選択されたタブのビューコントローラーがUINavigationControllerであることを確認
-    guard let navController = tabBarController.selectedViewController as? UINavigationController else {
-      print("Error: Selected view controller is not UINavigationController")
-      return
-    }
-    
-    // ナビゲーションスタックのルートまで戻る（途中の画面をクリア）
-    navController.popToRootViewController(animated: false)
-    
-    // トップのビューコントローラーがUIPageViewControllerであることを確認
-    guard let pageViewController = navController.topViewController as? UIPageViewController else {
-      print("Error: Top view controller is not UIPageViewController")
-      return
-    }
-    
-    // 新しいTopViewControllerをインスタンス化して表示
-    let topVC = TopViewController()
-    pageViewController.setViewControllers([topVC],
-                                          direction: .forward,
-                                          animated: false)
-  } else {
-    print("Error: Could not find window scene")
-  }
-}
-
-
